Download Source Data From ECB Europa Website
This commit introduces the ExchangeRateDownloader Trait which is used by the
ECBExchangeRateDownloader Class in order to download the source data from
https://www.ecb.europa.eu/stats/policy_and_exchange_rates/euro_reference_exchange_rates/html/index.en.html

The download() function parses the HTML document and fetches the downloadable link for the zip file.

When filePath (local file path for the zip file) is not provided in the ExchangeRateService API call,
it will automatically retrieve the zip file from the link mentioned above.